Fix issue of teapot being seen through table
For part 4 of of the worksheet, the reflection is fixed to no longer show for the portion that is under the table, but the real teapot can still be seen. The fix draws the teapot after the table and uses the depth test to ensure that parts of the teapot under the table aren't drawn.